compile fix

+++ ceammc/ext/src/lib/ceammc_xdatastorage.cpp
@@ -91,7 +91,7 @@ size_t XDataStorage::refCount(const DataDesc& desc)
 
 DataId XDataStorage::generateId(const AbstractData* data)
 {
-    size_t hash = boost::hash_value<size_t>(reinterpret_cast<size_t>(data));
+    size_t hash = boost::hash_value<long>(reinterpret_cast<long>(data));
     boost::hash_combine(hash, data->type());
     // NB: data type truncation!
     Atom a(DataDesc(0, DataId(hash)));
